refactoring gql folder structure
