+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "start": "nodemon app.js",
     "test": "echo \"Error: no test specified\" && exit 1",
-    "new": "nodemon src/app.js --exec babel-node"
+    "new": "nodemon src/app.js --ext js,graphql --exec babel-node"
   },
   "repository": {
     "type": "git",
