+++ src/schema.graphql
@@ -1,0 +1,66 @@
+type Query {
+  users(query: String): [User!]!
+  playgrounds(query: String): [Playground!]!
+  reviews(query: String): [Review!]!
+}
+
+type Mutation {
+  createUser(data: CreateUserInput): User!
+  deleteUser(id: ID!): User!
+  createPlayground(data: CreatePlaygroundInput): Playground!
+  deletePlayground(id: ID!): Playground!
+  createReview(data: CreateReviewInput): Review!
+  deleteReview(id: ID!): Review!
+}
+
+input CreateUserInput {
+  email: String!
+  username: String!
+}
+
+input CreatePlaygroundInput {
+  name: String!
+  image: String
+  price: Float
+  description: String!
+  location: String!
+  author: ID!
+}
+
+input CreateReviewInput {
+  body: String!
+  rating: Float!
+}
+
+type User {
+  id: ID!
+  email: String!
+  username: String!
+  playgrounds: [Playground!]!
+  reviews: [Review!]!
+}
+
+type Playground {
+  id: ID!
+  name: String!
+  image: String
+  geometry: Geometry
+  price: Float
+  description: String!
+  location: String!
+  author: User!
+  reviews: [Review!]!
+}
+
+type Review {
+  id: ID!
+  body: String!
+  rating: Float!
+  author: User!
+  playground: Playground!
+}
+
+type Geometry {
+  type: String!
+  coordinates: [Float!]
+}
